refactor(ui): Kurulum 9 adim bilesenini ayri dosyaya cikar (Faz E)
screens/KurulumSihirbazi/: adimlar.tsx (9 adim + InfoKutu, ~810 satir) + tipler.ts
(BildirimAyarlari/MuhafizYogunluk/KonumDurumu). Adimlar zaten props-tabanli modul
bilesenleriydi -> temiz tasima. KurulumSihirbaziSayfasi 1289 -> 488 satir
(stiller+adimlar+tipler ile toplam 1568 -> 488, ~%69 azalma). Orphan importlar
temizlendi.

388 test geciyor, 0 tip hatasi.
